# Conflicts:
#	Classes/Configuration/ExtensionConfiguration.php
#	Classes/Controller/Backend/ApiCoreController.php
#	Classes/Mapper/TcaMapper.php
#	Classes/Service/ApiTypoScriptSetupService.php
#	Classes/Service/PathAnalysisService.php
#	Classes/Service/Router.php
#	README.md
#	composer.json
#	docs/APIs.md
#	ext_conf_template.txt
#	ext_emconf.php
#	tests/Unit/Mapper/TcaMapperTest.php
#	tests/Unit/Middleware/ApiRequestMiddlewareTest.php
#	tests/Unit/Service/EndpointDiscoveryServiceTest.php
#	tests/Unit/Service/OpenApiServiceTest.php
#	tests/Unit/Service/RouterTest.php
#	tests/Unit/Service/UserAuthServiceTest.php

+++ Classes/Controller/TestController.php
@@ -282,7 +282,7 @@ class TestController {
 		$body = \is_array($body) ? $body : [];
 
 		$item = [
-			'id' => ((int) time() % 100000),
+			'id' => (time() % 100000),
 			'name' => (string) ($body['name'] ?? ''),
 			'type' => (string) ($body['type'] ?? ''),
 			'isDraft' => $this->isTruthy($body['isDraft'] ?? FALSE),
